# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Credit Broker 2.0",
        "description": "Документация\n * Во всех ответах приходит объект: `data`, кроме ошибок: В ошибках `401` имеется только поле `message`; в ошибке `422` имеются поля: `errors и message`.\n * Для авторизованных запросов необходимо вставить в HEADER `Authorization ` Bearer токен полученный при авторизации. Пример:\n *     Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbM3MjlkIn0.eyJhdWQiOiI2IiwianRpIjoiYThlYjY3NGVjMzIyMmU3MWY5Nzk3MzMwMjU\n * Токен можно получить при входе в систему или регистрации. Сам токен будет расположен в ответе HEADER:\n *      X-Access-Token: eyJ0eXAiOiJKV1QiLCJhbM3MjlkIn0.eyJhdWQiOiI2IiwianRpIjoiYThlYjY3NGVjMzIyMmU3MWY5Nzk3MzMwMjU\n *      X-Access-Token-Expires-At: 2023-11-08 10:13:18\n *  Важно! На всех запросах, кроме DEV - необходимо отправлять в HEADER Platform-Code и Device-Id. Иначе будет приходить  `403 ошибка`\n *      Platform-Code: ios\n *      Device-Id: DFIE3232dD\n *\n * На всех запросах в HEADER ОБЯЗАТЕЛЬНО установить:\n *     Accept: application/json",
        "contact": {
            "email": "develop@tsay.uz"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API LOCAL SERVER"
        },
        {
            "url": "http://{tenant}.localhost",
            "description": "API LOCAL SERVER WITH TENANT",
            "variables": {
                "tenant": {
...
            }
...
    "paths": {
...
        "/v1/profile/users": {
            "get": {
                "tags": [
                    "Центральный пользователь"
                ],
                "summary": "Получение всех центральных пользователей системы",
                "operationId": "ff717a96033dab36fb91d1460a2432f8",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Profile.UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
        "/v1/profile/users/{id}": {
            "get": {
                "tags": [
                    "Центральный пользователь"
                ],
                "summary": "Получение информации о центральном пользователе системы",
                "operationId": "10fc2b60117edf435e837bcd8cb4cfd0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Profile.UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "put": {
                "tags": [
                    "Центральный пользователь"
                ],
                "summary": "Обновление информации о центральном пользователе системы",
                "operationId": "409d3a199f1967b1790659be5d1099be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Тело запроса для обновления данных пользователя",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "first_name": "Vasya",
                                "last_name": "Pupkin"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Profile.UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Центральный пользователь"
                ],
                "summary": "Удаление центрального пользователя",
                "operationId": "78ac17e1766a17b95af3ddd027f669dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Profile.UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "v1.Profile.UserResource": {
                "description": "Class UserResource.",
                "properties": {
                    "id": {
                        "description": "ID пользователя",
                        "type": "integer",
                        "example": ""
                    },
                    "phone": {
                        "description": "Номер телефона пользователя",
                        "type": "string",
                        "example": ""
                    },
                    "first_name": {
                        "description": "Имя",
                        "type": "string",
                        "example": ""
                    },
                    "last_name": {
                        "description": "Фамилия",
                        "type": "string",
                        "example": ""
                    },
                    "is_active": {
                        "description": "Статус пользователя",
                        "type": "boolean",
                        "example": ""
                    },
                    "created_at": {
                        "description": "Дата создания сотрудника",
                        "type": "string",
                        "format": "date-time",
                        "example": ""
                    },
                    "updated_at": {
                        "description": "Дата обновления сотрудника",
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Пример: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImp0aSI6IjI0OT",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            },
            "Platform-Code": {
                "type": "apiKey",
                "description": "Код платформы",
                "name": "Platform-Code",
                "in": "header"
            },
            "Device-Id": {
                "type": "apiKey",
                "description": "ID устройства",
                "name": "Device-Id",
                "in": "header"
            },
            "Company-slug": {
                "type": "apiKey",
                "description": "Slug компании",
                "name": "Company",
                "in": "header"
            }
        }
    }
}